Modified the user DAO to use a file to store information about users
